requirements.txt
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.5
cssselect==1.1.0
dateparser==1.0.0
Django==3.2.7
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-environ==0.7.0
django-grappelli==2.15.1
facebook-scraper==0.2.46
fake-useragent==0.1.11
idna==3.2
importlib-metadata==4.8.1
lxml==4.6.3
parse==1.19.0
pyee==8.2.2
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tqdm==4.62.2
tzlocal==3.0
urllib3==1.26.6
w3lib==1.22.0
websockets==9.1
zipp==3.5.0
